Rename "resource_listwatch{ => error}"
